feat: further improvements to tables and policies for deletion

+++ app/Http/Controllers/WorkspaceUserController.php
@@ -6,6 +6,7 @@ namespace App\Http\Controllers;
 
 use App\Actions\WorkspaceUser\StoreWorkspaceUser;
 use App\Actions\WorkspaceUser\UpdateWorkspaceUser;
+use App\Data\Request\WorkspaceUser\WorkspaceUserIndexData;
 use App\Data\Request\WorkspaceUser\WorkspaceUserStoreData;
 use App\Data\Request\WorkspaceUser\WorkspaceUserUpdateData;
 use App\Data\Resource\WorkspaceUser\WorkspaceUserResource;
@@ -15,10 +16,34 @@ use App\Models\Workspace;
 use App\Models\WorkspaceUser;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Gate;
+use Spatie\LaravelData\Optional;
+use Spatie\LaravelData\PaginatedDataCollection;
 use Symfony\Component\HttpFoundation\Response;
 
 final class WorkspaceUserController extends Controller
 {
+    public function index(WorkspaceUserIndexData $data): JsonResponse
+    {
+        /** @var User $user */
+        $user = auth()->user();
+
+        $workspaceUsers = WorkspaceUser::query()
+            ->with([
+                'user',
+                'workspace',
+            ])
+            ->whereWorkspaceId($user->active_workspace_id)
+            ->paginate(
+                perPage: ! $data->per_page instanceof Optional ? $data->per_page : 10,
+                page: ! $data->page instanceof Optional ? $data->page : 1,
+            );
+
+        return $this->success(
+            WorkspaceUserResource::collect($workspaceUsers, PaginatedDataCollection::class),
+            'Workspace users retrieved successfully',
+        );
+    }
+
     public function store(Workspace $workspace, WorkspaceUserStoreData $data): JsonResponse
     {
         Gate::allowIf(fn (User $user) => $user->id === $workspace->owner_id && ! $workspace->is_default);
@@ -57,12 +82,12 @@ final class WorkspaceUserController extends Controller
 
     public function destroy(Workspace $workspace, WorkspaceUser $workspaceUser): JsonResponse
     {
-        Gate::allowIf(fn (User $user) => $user->id === $workspaceUser->user_id && ! $workspace->is_default);
+        Gate::authorize('delete', [$workspaceUser, $workspace]);
 
         $workspaceUser->delete();
 
         return $this->success(
-            message: "You have left the workspace {$workspace->name}.",
+            message: "User has left the workspace {$workspace->name}.",
             code: Response::HTTP_NO_CONTENT,
         );
     }
